backend/requirements.txt
# Indeed Job Ingestion Requirements

# HTTP client
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# PostgreSQL driver
psycopg2-binary>=2.9.9

# MySQL driver (optional, if using MySQL)
mysql-connector-python>=8.2.0
